Refactor worktree create request parsing
